requirements.txt
requests
beautifulsoup4
flask
flask-cors
python-dotenv
